Handle field id for parquet protobuf
When read back a parquet file with protobuf schema, try to match
the parquet schema to protobuf with respect to field id.
If the parquet file is written with a previsous parquet-protobuf (
prior to 1.9.1), field id is not persisted in parquet file metadata.
When reading these legacy file, schema converter falls back to
matching fields by name.

Change-Id: If2e2d52405bb625d2662f29f79b0a496ba6f0652
JIRA: PARQUET-951