Make ProtoReadSupport work the same as ProtoMessageConverter
If an unknown enum value is serialized, it will default to an unknown
EnumValueDescriptor with name as:
"UNKNOWN_ENUM_VALUE_" + parent.getName() + "_" + number
But when reading the value from parquet, it expects correct enum
values.
It's the same problem with added enum values and the protobuf schema
not updated yet on reading side.

Change-Id: I11d0676878c128aee6cefceb444996f56dbbe905