Patch to fix "unknown enum" issue
Protobuf can set enum field using number, while a number does not
match any enum value defined in the schema, it is still accepted
and a label "UNKNOWN_ENUM_<enumName>_<number>" is generated when
we use protobuf reflection API (proto descriptors) to access it.
And in parquet-protobuf, we rely on protobuf reflection API to
convert forward/backward between the two world.

There are two cases of unknown enum while using parquet-protobuf:
  1. Protobuf already contains unknown enum when we write it to
  parquet (sometimes people manipulate enum using numbers), so it
  will write a label "UNKNOWN_ENUM_*" as string in parquet. And
  when we read it back to protobuf, we found this "true" unknown
  value
  2. Protobuf contains valid value when write to parquet, but the
  reader uses an outdated proto schema which misses some enum
  values. So the not-in-old-schema enum values are "unknown" to the
  reader.

Previous behavior of parquet-proto reader is to reject in both
cases with some runtime exception.

To be able to handle the problems:
We keep enum (name -> number) mapping in the parquet metadata, so
that in read time, reader can discover the number and use protobuf
reflection API to set enum number.
Keep in mind though, for the case reading enum with outdated schema
(case 2), the enum read back will have the right number, but the
label is set to "UNKNOW_ENUM_*" with the number as part of it. So
this feature is helpful only if the user is using number to
manipulate enum data.
And for old data containing "true" unknown value (thus case 1)
created before this patch, we now try to parse the string regarding
to the "UNKNOWN_ENUM_*" pattern.
If we read old data created before this patch (thus name -> number
is not available), with an outdated schema, and we find some enum
value not defined in the schema nor following "UNKNOWN_ENUM_*"
pattern, we could either fail the job by raising an exception or
treat the value as unknown enum with number -1, by setting a flag
in the configuration.