Revert "Add all known enum values to the metadata"
This reverts commit be6e31d669f18412bbee1a1cb70f958f9dae5540. (with necessary rebase)

The commit was to address the problem while merging/consolidating parquet (proto)
files that contain different subsets of enum value in a enum field, and thus contain
different value in the metadata.

The consolidation problem itself is now solved in the ParquetConsolidatorInputFormat
(in datadisco), which is now able to combine subsets of enum values (known or unknown)
in parquet metadata. So the commit is no more necessary.

Change-Id: Ic60b128c0950d9dc71897980f92838ae2f4acb88